tested for clashing local variables which passed without any fixes. Cleaned up internals and renamed/moved around adapters. At this point updating the default class loader is also not needed.
