+++ mixin-base-test/src/test/java/net/yakclient/apitests/RegistryTest.java
@@ -6,20 +6,17 @@ import net.yakclient.mixin.base.registry.MixinRegistry;
 import net.yakclient.mixin.base.registry.RegistryConfigurator;
 import org.junit.jupiter.api.Test;
 
-import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 
 public class RegistryTest {
-    //for runtime -Djava.system.class.loader=net.yakclient.mixin.internal.loader.ProxyClassLoader
-
-    public static void main(String[] args) throws ClassNotFoundException, NoSuchMethodException, IllegalAccessException, InvocationTargetException, InstantiationException, IOException {
+    public static void main(String[] args) throws ClassNotFoundException, NoSuchMethodException, IllegalAccessException, InvocationTargetException, InstantiationException {
 
 
         final RegistryConfigurator configure = RegistryConfigurator.configure().addSafePackage("net.questcraft").addTarget("net.questcraft");
         final MixinRegistry mixinRegistry = configure.create();
         {
-            mixinRegistry.registerMixin(MixinTestCase.class, Runnable::run).dumpAll();
+            mixinRegistry.registerMixin(MixinTestCase.class).dumpAll();
 
             final Class<?> aClass = mixinRegistry.retrieveClass(net.questcraft.apitests.MixinSourceClassTest.class.getName()); //target.retrieveClass(MixinSourceClassTest.class.getName());
             final Constructor<?> constructor = aClass.getConstructor(String.class);
@@ -29,7 +26,7 @@ public class RegistryTest {
         }
 
         {
-            mixinRegistry.registerMixin(net.questcraft.apitests.SecondMixinTestCase.class, Runnable::run).dumpAll();
+            mixinRegistry.registerMixin(net.questcraft.apitests.SecondMixinTestCase.class).dumpAll();
