+++ mixin-base/src/main/java/net/yakclient/mixin/base/internal/instruction/DirectInsnInterceptor.java
@@ -3,8 +3,8 @@ package net.yakclient.mixin.base.internal.instruction;
 import net.yakclient.mixin.base.YakMixins;
 import org.objectweb.asm.tree.MethodNode;
 
-public class ASMInsnInterceptor extends MethodNode implements InstructionInterceptor {
-    public ASMInsnInterceptor() {
+public class DirectInsnInterceptor extends MethodNode implements InstructionInterceptor {
+    public DirectInsnInterceptor() {
         super(YakMixins.ASM_VERSION);
     }
