+++ mixin-base/src/main/java/net/yakclient/mixin/base/internal/instruction/ProxyInsnInterceptor.java
@@ -9,10 +9,10 @@ import java.util.UUID;
 
 import static org.objectweb.asm.Opcodes.*;
 
-public class ProxyASMInsnInterceptor extends ASMInsnInterceptor {
+public class ProxyInsnInterceptor extends DirectInsnInterceptor {
     private final UUID pointer;
 
-    public ProxyASMInsnInterceptor(UUID pointer) {
+    public ProxyInsnInterceptor(UUID pointer) {
         this.pointer = pointer;
     }
 
@@ -29,45 +29,24 @@ public class ProxyASMInsnInterceptor extends ASMInsnInterceptor {
 
         insn.add(l0);
 
-//        insn.add(new FieldInsnNode(GETSTATIC, "java/lang/System", "out", "Ljava/io/PrintStream;"));
-//        insn.add(new LdcInsnNode("Byte code printed"));
-//        insn.add(new MethodInsnNode(INVOKEVIRTUAL, "java/io/PrintStream", "println", "(Ljava/lang/String;)V"));
-
         insn.add(new TypeInsnNode(NEW, "java/util/UUID"));
         insn.add(new InsnNode(DUP));
         insn.add(new LdcInsnNode(this.pointer.getMostSignificantBits()));
         insn.add(new LdcInsnNode(this.pointer.getLeastSignificantBits()));
         insn.add(new MethodInsnNode(INVOKESPECIAL, "java/util/UUID", "<init>", "(JJ)V", false));
-//        insn.add(new VarInsnNode(ASTORE, 2));
-//
-//        // GETSTATIC java/lang/System.out : Ljava/io/PrintStream;
-//        //    LDC "ok"
-//        //    INVOKEVIRTUAL java/io/PrintStream.println (Ljava/lang/String;)V
 //
         final String proxyResName = this.getRuntimeName(FunctionalProxy.ProxyResponseData.class);
 
         insn.add(l1);
-//        insn.add(new VarInsnNode(ALOAD, 2));
         insn.add(new MethodInsnNode(INVOKESTATIC, this.getRuntimeName(MixinProxyManager.class), "proxy", "(Ljava/util/UUID;)L" + proxyResName + ";", false));
-//        insn.add(new VarInsnNode(ASTORE, 3));
 
         insn.add(l2);
-//        insn.add(new VarInsnNode(ALOAD, 3));
-        //"net/yakclient/mixin/registry/proxy/MixinProxyManager"
-        //"net/yakclient/mixin/registry/FunctionalProxy$ProxyResponseData"
         insn.add(new FieldInsnNode(GETFIELD, proxyResName, "cancel", "Z"));
-//        insn.add(new InsnNode(DUP));
-
-//        insn.add(new FieldInsnNode(GETSTATIC, "java/lang/System", "out", "Ljava/io/PrintStream;"));
-//        insn.add(new FieldInsnNode(GETFIELD, proxyResName, "cancel", "Z"));
-
-//        insn.add(new MethodInsnNode(INVOKEVIRTUAL, "java/io/PrintStream", "println", "(Ljava/lang/String;)V"));
 
         insn.add(new JumpInsnNode(IFNE, l3));
+
         insn.add(l4);
-//
         insn.add(super.intercept().getInstructions());
-//
         insn.add(l3);
 
         return new DirectInstruction(insn);
