--- mixin-base/src/main/java/net/yakclient/mixin/base/internal/instruction/adapter/LocalsUpdateAdapter.java
@@ -1,4 +1,0 @@
-package net.yakclient.mixin.base.internal.instruction.adapter;
-
-public class LocalsUpdateAdapter  {
-}
